requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
PyPDF2>=3.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
python-multipart>=0.0.6
requests>=2.31.0
pytest>=7.4.0
httpx>=0.25.0
google-generativeai>=0.3.0
sqlalchemy>=2.0.23
databases[aiosqlite]>=0.8.0
aiosqlite>=0.19.0
alembic>=1.12.1
slowapi>=0.1.9
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0,<2.0.0
PyJWT>=2.8.0
bcrypt>=4.1.2
scikit-learn>=1.3.0
psutil>=5.9.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
redis>=5.0.0
aioredis>=2.0.0
schedule>=1.2.0
asyncio-throttle>=1.0.2
prometheus-client>=0.19.0
structlog>=23.2.0
colorlog>=6.8.0
email-validator>=2.1.0
protobuf==3.20.3
